tests/wpt/tests/html/canvas/offscreen/shadows/2d.shadow.offset.negativeY.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.shadow.offset.negativeY</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.shadow.offset.negativeY</h1>
<p class="desc">Shadows can be offset with negative y</p>


<script>
var t = async_test("Shadows can be offset with negative y");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.fillStyle = '#0f0';
  ctx.shadowColor = '#0f0';
  ctx.shadowOffsetY = -25;
  ctx.fillRect(0, 25, 100, 25);
  _assertPixel(canvas, 50,12, 0,255,0,255);
  _assertPixel(canvas, 50,37, 0,255,0,255);
  t.done();

});
</script>
